deepdream/requirements.txt
imageio==2.1.2
numpy==1.12.1
Pillow==3.4.2
tensorflow-gpu==1.0.1
